Peer Review
- Rename library from `diskobject` to `vegadisk`
- Rename namespace from `vegastrike::diskobject` to `vegadisk`
- Move posh to its own static library; we really don't modify it
