+++ engine/src/aldrv/al_sound.cpp
@@ -34,7 +34,7 @@
 #include "cmd/unit_generic.h"
 #include "gfx_generic/cockpit_generic.h"
 #include "root_generic/options.h"
-#include "root_generic/posh.h"
+#include "posh/posh.h"
 #include "src/universe.h"
 
 #ifdef HAVE_AL
