+++ libraries/posh/posh.cpp
@@ -33,7 +33,7 @@
  *  @date    2002
  *  @brief   Portable Open Source Harness primary source file
  */
-#include "root_generic/posh.h"
+#include "posh/posh.h"
 
 #if !defined POSH_NO_FLOAT
 #  define POSH_FLOAT_STRING "enabled"
