+++ libraries/root_generic/macosx_math.cpp
@@ -28,7 +28,7 @@
 
 #include "root_generic/macosx_math.h"
 #include <string>
-#include "root_generic/posh.h"
+#include "posh/posh.h"
 #if defined (__APPLE__)
 //these stuffs are included in OSX 10.4 and above--so just check for x86
 #include <stdio.h>
