+++ libraries/vegadisk/CMakeLists.txt
@@ -28,27 +28,27 @@
 
 MESSAGE(STATUS "PROCESSING DIRECTORY: ${CMAKE_CURRENT_SOURCE_DIR}")
 
-ADD_LIBRARY(vegastrike_disk_object STATIC
+ADD_LIBRARY(vegastrike_vegadisk STATIC
     diskobject.h
     diskobject.cpp
 )
 
 #TARGET_COMPILE_FEATURES(vegastrike_gfx_generic PUBLIC cxx_std_11)
-SET_PROPERTY(TARGET vegastrike_disk_object PROPERTY CXX_STANDARD 14)
-SET_PROPERTY(TARGET vegastrike_disk_object PROPERTY CXX_STANDARD_REQUIRED TRUE)
-SET_PROPERTY(TARGET vegastrike_disk_object PROPERTY CXX_EXTENSIONS ON)
-SET_PROPERTY(TARGET vegastrike_disk_object PROPERTY POSITION_INDEPENDENT_CODE TRUE)
+SET_PROPERTY(TARGET vegastrike_vegadisk PROPERTY CXX_STANDARD 14)
+SET_PROPERTY(TARGET vegastrike_vegadisk PROPERTY CXX_STANDARD_REQUIRED TRUE)
+SET_PROPERTY(TARGET vegastrike_vegadisk PROPERTY CXX_EXTENSIONS ON)
+SET_PROPERTY(TARGET vegastrike_vegadisk PROPERTY POSITION_INDEPENDENT_CODE TRUE)
 
-TARGET_COMPILE_DEFINITIONS(vegastrike_disk_object PUBLIC "BOOST_ALL_DYN_LINK" "$<$<CONFIG:Debug>:BOOST_DEBUG_PYTHON>")
+TARGET_COMPILE_DEFINITIONS(vegastrike_vegadisk PUBLIC "BOOST_ALL_DYN_LINK" "$<$<CONFIG:Debug>:BOOST_DEBUG_PYTHON>")
 IF (WIN32)
-    TARGET_COMPILE_DEFINITIONS(vegastrike_disk_object PUBLIC BOOST_USE_WINAPI_VERSION=0x0A00)
-    TARGET_COMPILE_DEFINITIONS(vegastrike_disk_object PUBLIC _WIN32_WINNT=0x0A00)
-    TARGET_COMPILE_DEFINITIONS(vegastrike_disk_object PUBLIC WINVER=0x0A00)
-    TARGET_COMPILE_DEFINITIONS(vegastrike_disk_object PUBLIC "$<$<CONFIG:Debug>:Py_DEBUG>")
+    TARGET_COMPILE_DEFINITIONS(vegastrike_vegadisk PUBLIC BOOST_USE_WINAPI_VERSION=0x0A00)
+    TARGET_COMPILE_DEFINITIONS(vegastrike_vegadisk PUBLIC _WIN32_WINNT=0x0A00)
+    TARGET_COMPILE_DEFINITIONS(vegastrike_vegadisk PUBLIC WINVER=0x0A00)
+    TARGET_COMPILE_DEFINITIONS(vegastrike_vegadisk PUBLIC "$<$<CONFIG:Debug>:Py_DEBUG>")
 ENDIF()
 
-TARGET_INCLUDE_DIRECTORIES(vegastrike_disk_object SYSTEM PRIVATE ${TST_INCLUDES})
-TARGET_INCLUDE_DIRECTORIES(vegastrike_disk_object PRIVATE
+TARGET_INCLUDE_DIRECTORIES(vegastrike_vegadisk SYSTEM PRIVATE ${TST_INCLUDES})
+TARGET_INCLUDE_DIRECTORIES(vegastrike_vegadisk PRIVATE
         # VS engine headers
         ${Vega_Strike_SOURCE_DIR}
         ${Vega_Strike_SOURCE_DIR}/engine
@@ -62,7 +62,7 @@ TARGET_INCLUDE_DIRECTORIES(vegastrike_disk_object PRIVATE
         ${Vega_Strike_BINARY_DIR}/engine/src
 )
 IF (NEED_LINKING_AGAINST_LIBM)
-    TARGET_LINK_LIBRARIES(vegastrike_disk_object PRIVATE m)
+    TARGET_LINK_LIBRARIES(vegastrike_vegadisk PRIVATE m)
 ENDIF()
-TARGET_LINK_LIBRARIES(vegastrike_disk_object PRIVATE ${TST_LIBS} GLUT::GLUT)
-TARGET_LINK_LIBRARIES(vegastrike_disk_object PUBLIC vegastrike_root_generic)
+TARGET_LINK_LIBRARIES(vegastrike_vegadisk PRIVATE ${TST_LIBS} GLUT::GLUT)
+TARGET_LINK_LIBRARIES(vegastrike_vegadisk PUBLIC vegastrike_root_generic)
